+++ packages/client/src/fs/index.ts
@@ -3,5 +3,5 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-export const LANGUAGE_ID = 'cpp';
-export const WORKSPACE_PATH = '/home/web_user';
+export * from './definitions.js';
+export * from './endpoints/defaultEndpoint.js';