+++ packages/examples/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this npm module are documented in this file.
 
+## [2024.10.5] - 2024-10-2x
+
+- Added clangd example.
+
 ## [2024.10.4] - 2024-10-23
 
 - Updated to `monaco-languageclient@9.0.0-next.5`, `monaco-editor-wrapper@6.0.0-next.5` and `@typefox/monaco-editor-react@6.0.0-next.5`.